tests/wpt/web-platform-tests/webmessaging/message-channels/001.html

<!doctype html>
<title>basic messagechannel test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t) {
  var channel = new MessageChannel();
  channel.port1.postMessage(1);
  channel.port2.onmessage = t.step_func(
    function(e) {
      assert_equals(e.data, 1);
      t.done();
    });
  channel.port2.start();
});
</script>
